requirements.txt
notion-client
python-dotenv
google-generativeai
